[Swagger] File-scoped namespaces
Switch to file-scoped namespaces for `Swashbuckle.AspNetCore.Swagger*`.
