+++ src/Swashbuckle.AspNetCore.Swagger/ISwaggerDocumentSerializer.cs
@@ -2,19 +2,18 @@
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Writers;
 
-namespace Swashbuckle.AspNetCore.Swagger
+namespace Swashbuckle.AspNetCore.Swagger;
+
+/// <summary>
+/// Provide an implementation for this interface if you wish to customize how the OpenAPI document is written.
+/// </summary>
+public interface ISwaggerDocumentSerializer
 {
     /// <summary>
-    /// Provide an implementation for this interface if you wish to customize how the OpenAPI document is written.
+    /// Serializes an OpenAPI document.
     /// </summary>
-    public interface ISwaggerDocumentSerializer
-    {
-        /// <summary>
-        /// Serializes an OpenAPI document.
-        /// </summary>
-        /// <param name="document">The OpenAPI document that should be serialized.</param>
-        /// <param name="writer">The writer to which the document needs to be written.</param>
-        /// <param name="specVersion">The OpenAPI specification version to serialize as.</param>
-		void SerializeDocument(OpenApiDocument document, IOpenApiWriter writer, OpenApiSpecVersion specVersion);
-	}
+    /// <param name="document">The OpenAPI document that should be serialized.</param>
+    /// <param name="writer">The writer to which the document needs to be written.</param>
+    /// <param name="specVersion">The OpenAPI specification version to serialize as.</param>
+    void SerializeDocument(OpenApiDocument document, IOpenApiWriter writer, OpenApiSpecVersion specVersion);
 }
