+++ src/Swashbuckle.AspNetCore.Swagger/SwaggerMiddleware.cs
@@ -10,188 +10,187 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Writers;
 
-namespace Swashbuckle.AspNetCore.Swagger
+namespace Swashbuckle.AspNetCore.Swagger;
+
+internal sealed class SwaggerMiddleware
 {
-    internal sealed class SwaggerMiddleware
+    private static readonly Encoding UTF8WithoutBom = new UTF8Encoding(false);
+
+    private readonly RequestDelegate _next;
+    private readonly SwaggerOptions _options;
+    private readonly TemplateMatcher _requestMatcher;
+#if !NETSTANDARD
+    private readonly TemplateBinder _templateBinder;
+#endif
+
+    public SwaggerMiddleware(
+        RequestDelegate next,
+        SwaggerOptions options)
     {
-        private static readonly Encoding UTF8WithoutBom = new UTF8Encoding(false);
-
-        private readonly RequestDelegate _next;
-        private readonly SwaggerOptions _options;
-        private readonly TemplateMatcher _requestMatcher;
-#if !NETSTANDARD
-        private readonly TemplateBinder _templateBinder;
-#endif
-
-        public SwaggerMiddleware(
-            RequestDelegate next,
-            SwaggerOptions options)
-        {
-            _next = next;
-            _options = options ?? new SwaggerOptions();
-            _requestMatcher = new TemplateMatcher(TemplateParser.Parse(_options.RouteTemplate), []);
-        }
+        _next = next;
+        _options = options ?? new SwaggerOptions();
+        _requestMatcher = new TemplateMatcher(TemplateParser.Parse(_options.RouteTemplate), []);
+    }
 
 #if !NETSTANDARD
-        [ActivatorUtilitiesConstructor]
-        public SwaggerMiddleware(
-            RequestDelegate next,
-            SwaggerOptions options,
-            TemplateBinderFactory templateBinderFactory) : this(next, options)
-        {
-            _templateBinder = templateBinderFactory.Create(RoutePatternFactory.Parse(_options.RouteTemplate));
-        }
+    [ActivatorUtilitiesConstructor]
+    public SwaggerMiddleware(
+        RequestDelegate next,
+        SwaggerOptions options,
+        TemplateBinderFactory templateBinderFactory) : this(next, options)
+    {
+        _templateBinder = templateBinderFactory.Create(RoutePatternFactory.Parse(_options.RouteTemplate));
+    }
 #endif
 
-        public async Task Invoke(HttpContext httpContext, ISwaggerProvider swaggerProvider)
+    public async Task Invoke(HttpContext httpContext, ISwaggerProvider swaggerProvider)
+    {
+        if (!RequestingSwaggerDocument(httpContext.Request, out string documentName, out string extension))
         {
-            if (!RequestingSwaggerDocument(httpContext.Request, out string documentName, out string extension))
-            {
-                await _next(httpContext);
-                return;
-            }
-
-            try
-            {
-                var basePath = httpContext.Request.PathBase.HasValue
-                    ? httpContext.Request.PathBase.Value
-                    : null;
-
-                OpenApiDocument swagger;
-                var asyncSwaggerProvider = httpContext.RequestServices.GetService<IAsyncSwaggerProvider>();
-
-                if (asyncSwaggerProvider is not null)
-                {
-                    swagger = await asyncSwaggerProvider.GetSwaggerAsync(
-                        documentName: documentName,
-                        host: null,
-                        basePath: basePath);
-                }
-                else
-                {
-                    swagger = swaggerProvider.GetSwagger(
-                        documentName: documentName,
-                        host: null,
-                        basePath: basePath);
-                }
-
-                // One last opportunity to modify the Swagger Document - this time with request context
-                foreach (var filter in _options.PreSerializeFilters)
-                {
-                    filter(swagger, httpContext.Request);
-                }
-
-                if (extension is ".yaml" or ".yml")
-                {
-                    await RespondWithSwaggerYaml(httpContext.Response, swagger);
-                }
-                else
-                {
-                    await RespondWithSwaggerJson(httpContext.Response, swagger);
-                }
-            }
-            catch (UnknownSwaggerDocument)
-            {
-                httpContext.Response.StatusCode = 404;
-            }
+            await _next(httpContext);
+            return;
         }
 
-        private bool RequestingSwaggerDocument(HttpRequest request, out string documentName, out string extension)
+        try
         {
-            documentName = null;
-            extension = null;
+            var basePath = httpContext.Request.PathBase.HasValue
+                ? httpContext.Request.PathBase.Value
+                : null;
 
-            if (!HttpMethods.IsGet(request.Method))
+            OpenApiDocument swagger;
+            var asyncSwaggerProvider = httpContext.RequestServices.GetService<IAsyncSwaggerProvider>();
+
+            if (asyncSwaggerProvider is not null)
             {
-                return false;
-            }
-
-            var routeValues = new RouteValueDictionary();
-            if (_requestMatcher.TryMatch(request.Path, routeValues))
-            {
-#if !NETSTANDARD
-                if (_templateBinder != null && !_templateBinder.TryProcessConstraints(request.HttpContext, routeValues, out _, out _))
-                {
-                    return false;
-                }
-#endif
-                if (routeValues.TryGetValue("documentName", out var documentNameObject) && documentNameObject is string documentNameString)
-                {
-                    documentName = documentNameString;
-                    if (routeValues.TryGetValue("extension", out var extensionObject))
-                    {
-                        extension = $".{extensionObject}";
-                    }
-                    else
-                    {
-                        extension = Path.GetExtension(request.Path.Value);
-                    }
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        private async Task RespondWithSwaggerJson(HttpResponse response, OpenApiDocument swagger)
-        {
-            string json;
-
-            using (var textWriter = new StringWriter(CultureInfo.InvariantCulture))
-            {
-                var openApiWriter = new OpenApiJsonWriter(textWriter);
-
-                SerializeDocument(swagger, openApiWriter);
-
-                json = textWriter.ToString();
-            }
-
-            response.StatusCode = 200;
-            response.ContentType = "application/json;charset=utf-8";
-
-            await response.WriteAsync(json, UTF8WithoutBom);
-        }
-
-        private async Task RespondWithSwaggerYaml(HttpResponse response, OpenApiDocument swagger)
-        {
-            string yaml;
-
-            using (var textWriter = new StringWriter(CultureInfo.InvariantCulture))
-            {
-                var openApiWriter = new OpenApiYamlWriter(textWriter);
-
-                SerializeDocument(swagger, openApiWriter);
-
-                yaml = textWriter.ToString();
-            }
-
-            response.StatusCode = 200;
-            response.ContentType = "text/yaml;charset=utf-8";
-
-            await response.WriteAsync(yaml, UTF8WithoutBom);
-        }
-
-        private void SerializeDocument(
-            OpenApiDocument document,
-            IOpenApiWriter writer)
-        {
-            if (_options.CustomDocumentSerializer != null)
-            {
-                _options.CustomDocumentSerializer.SerializeDocument(document, writer, _options.OpenApiVersion);
+                swagger = await asyncSwaggerProvider.GetSwaggerAsync(
+                    documentName: documentName,
+                    host: null,
+                    basePath: basePath);
             }
             else
             {
-                switch (_options.OpenApiVersion)
-                {
-                    case Microsoft.OpenApi.OpenApiSpecVersion.OpenApi2_0:
-                        document.SerializeAsV2(writer);
-                        break;
+                swagger = swaggerProvider.GetSwagger(
+                    documentName: documentName,
+                    host: null,
+                    basePath: basePath);
+            }
 
-                    case Microsoft.OpenApi.OpenApiSpecVersion.OpenApi3_0:
-                    default:
-                        document.SerializeAsV3(writer);
-                        break;
+            // One last opportunity to modify the Swagger Document - this time with request context
+            foreach (var filter in _options.PreSerializeFilters)
+            {
+                filter(swagger, httpContext.Request);
+            }
+
+            if (extension is ".yaml" or ".yml")
+            {
+                await RespondWithSwaggerYaml(httpContext.Response, swagger);
+            }
+            else
+            {
+                await RespondWithSwaggerJson(httpContext.Response, swagger);
+            }
+        }
+        catch (UnknownSwaggerDocument)
+        {
+            httpContext.Response.StatusCode = 404;
+        }
+    }
+
+    private bool RequestingSwaggerDocument(HttpRequest request, out string documentName, out string extension)
+    {
+        documentName = null;
+        extension = null;
+
+        if (!HttpMethods.IsGet(request.Method))
+        {
+            return false;
+        }
+
+        var routeValues = new RouteValueDictionary();
+        if (_requestMatcher.TryMatch(request.Path, routeValues))
+        {
+#if !NETSTANDARD
+            if (_templateBinder != null && !_templateBinder.TryProcessConstraints(request.HttpContext, routeValues, out _, out _))
+            {
+                return false;
+            }
+#endif
+            if (routeValues.TryGetValue("documentName", out var documentNameObject) && documentNameObject is string documentNameString)
+            {
+                documentName = documentNameString;
+                if (routeValues.TryGetValue("extension", out var extensionObject))
+                {
+                    extension = $".{extensionObject}";
                 }
+                else
+                {
+                    extension = Path.GetExtension(request.Path.Value);
+                }
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private async Task RespondWithSwaggerJson(HttpResponse response, OpenApiDocument swagger)
+    {
+        string json;
+
+        using (var textWriter = new StringWriter(CultureInfo.InvariantCulture))
+        {
+            var openApiWriter = new OpenApiJsonWriter(textWriter);
+
+            SerializeDocument(swagger, openApiWriter);
+
+            json = textWriter.ToString();
+        }
+
+        response.StatusCode = 200;
+        response.ContentType = "application/json;charset=utf-8";
+
+        await response.WriteAsync(json, UTF8WithoutBom);
+    }
+
+    private async Task RespondWithSwaggerYaml(HttpResponse response, OpenApiDocument swagger)
+    {
+        string yaml;
+
+        using (var textWriter = new StringWriter(CultureInfo.InvariantCulture))
+        {
+            var openApiWriter = new OpenApiYamlWriter(textWriter);
+
+            SerializeDocument(swagger, openApiWriter);
+
+            yaml = textWriter.ToString();
+        }
+
+        response.StatusCode = 200;
+        response.ContentType = "text/yaml;charset=utf-8";
+
+        await response.WriteAsync(yaml, UTF8WithoutBom);
+    }
+
+    private void SerializeDocument(
+        OpenApiDocument document,
+        IOpenApiWriter writer)
+    {
+        if (_options.CustomDocumentSerializer != null)
+        {
+            _options.CustomDocumentSerializer.SerializeDocument(document, writer, _options.OpenApiVersion);
+        }
+        else
+        {
+            switch (_options.OpenApiVersion)
+            {
+                case Microsoft.OpenApi.OpenApiSpecVersion.OpenApi2_0:
+                    document.SerializeAsV2(writer);
+                    break;
+
+                case Microsoft.OpenApi.OpenApiSpecVersion.OpenApi3_0:
+                default:
+                    document.SerializeAsV3(writer);
+                    break;
             }
         }
     }
